UPDATE: Got ads working in production, I think
