+++ app/screens/LogIn.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { StyleSheet, View, Button, Text, KeyboardAvoidingView, ScrollView, Platform } from 'react-native';
 import analytics from '@react-native-firebase/analytics';
-import { InterstitialAd, TestIds, AdEventType } from '@react-native-firebase/admob';
 import crashlytics from '@react-native-firebase/crashlytics';
 // ----------------------------------------------------------------------------
 import { useAuthUpdate } from '../hooks/useAuth'
@@ -14,50 +13,17 @@ import Spacer from '../components/Spacer';
 import { sizes } from '../config/constants';
 import * as yup from 'yup'
 
-// ============================================================================
-const adUnitId = __DEV__ ? TestIds.INTERSTITIAL : 'ca-app-pub-2742026173933447~7286128693';
-
-const interstitial = InterstitialAd.createForAdRequest(adUnitId, {
-  requestNonPersonalizedAdsOnly: true,
-  keywords: ['fitness', 'exercie'],
-});
-// ============================================================================
-
-
-
 function LogIn({ navigation }) {
   const { logIn } = useAuthUpdate();
-  const [loaded, setLoaded] = React.useState(false);
-
-
 
   React.useEffect(() => {
     crashlytics().log('LogIn mounted');
-// ############################################################################
-    const eventListener = interstitial.onAdEvent(type => {
-      if (type === AdEventType.LOADED) {
-        setLoaded(true);
-      }
-    });
-    // Start loading the interstitial straight away
-    interstitial.load();
-// ############################################################################
-    // Unsubscribe from events on unmount
-    return () => {
-      eventListener(); // Unsubscribe from ad event listener
-    };
   }, []);
-
-  // No advert ready to show yet //Warning: Causes content to not render when logging out
-  // if (!loaded) {
-  //   return null;
-  // }
 
   function Buttons() {
     return (
       <>
       <Button title="Log In" onPress={() => {
-        // interstitial.show();
         logIn()
       }} />
       <Button title="Sign Up" onPress={() => navigation.navigate("SignUp")} />
